feat: convert pytorch models to onnx to allow more backends (#46)
This PR uses onnx files instead of pytorch files and moves the backend
of m6a prediction to burn-rs instead of tch-rs.

This allows us to use other ML backends in addition to pytorch, which
means we can still run m6a prediction without the massive pytorch
libraries and get the same results at the cost of some performance.

I have also added tests for m6a predictions for each PacBio chemistry to
ensure that the number and quality of predictions are identical to
previous versions regardless of backend. Will also be a useful test in
the future.


I also add a lot of new functionality to pyft.

This will be the basis of ft v0.4.0 when merged.